// frontend/plugins/welcome/src/api/models/EntHospital.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntHospitalEdges,
    EntHospitalEdgesFromJSON,
    EntHospitalEdgesFromJSONTyped,
    EntHospitalEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntHospital
 */
export interface EntHospital {
    /**
     * HospitalName holds the value of the "HospitalName" field.
     * @type {string}
     * @memberof EntHospital
     */
    hospitalName?: string;
    /**
     * 
     * @type {EntHospitalEdges}
     * @memberof EntHospital
     */
    edges?: EntHospitalEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntHospital
     */
    id?: number;
}

export function EntHospitalFromJSON(json: any): EntHospital {
    return EntHospitalFromJSONTyped(json, false);
}

export function EntHospitalFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntHospital {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'hospitalName': !exists(json, 'HospitalName') ? undefined : json['HospitalName'],
        'edges': !exists(json, 'edges') ? undefined : EntHospitalEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntHospitalToJSON(value?: EntHospital | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'HospitalName': value.hospitalName,
        'edges': EntHospitalEdgesToJSON(value.edges),
        'id': value.id,
    };
}